// apps/frontend/src/api/generated/model/createEventBody.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Constancia Backend API
 * Backend API contract for the Constancia frontend and Discord bot.
 * OpenAPI spec version: 0.1.0
 */
import type { CreateEventBodyPipelineItem } from './createEventBodyPipelineItem';

export type CreateEventBody = {
  name: string;
  type: string;
  channelId: string;
  shortCircuit?: boolean;
  pipeline: CreateEventBodyPipelineItem[];
};